Make tests/report in docker and codecov outside
Codecov needs the .git directory, that's why I have to use it outsie.
Coverage xml somewhat needs to be runned where the tests where run, I'm
not sure why since the db is there, but that's what it is.
I leave the .env.codecov generation for possible future use in frontend.